#48 Seperate Physical and Embedded into their own packages.
