+++ src/Microsoft.AspNet.FileProviders.Common/Implementation/NoopTrigger.cs
