+++ src/Microsoft.AspNet.FileProviders.Common/project.json
@@ -1,0 +1,16 @@
+{
+    "version": "1.0.0-*",
+    "shared": "**/*.cs",
+    "dependencies": {
+        "Microsoft.AspNet.FileProviders.Interfaces": "1.0.0-*"
+    },
+    "frameworks": {
+        "net45": { },
+        "dnx451": { },
+        "dnxcore50": {
+            "dependencies": {
+                "System.Linq": "4.0.0-beta-*"
+            }
+        }
+    }
+}
