+++ src/Microsoft.AspNet.FileProviders.Embedded/EmbeddedFileProvider.cs
