+++ src/Microsoft.AspNet.FileProviders.Embedded/Properties/AssemblyInfo.cs
