+++ src/Microsoft.AspNet.FileProviders.Physical/FileChangeTrigger.cs
