+++ src/Microsoft.AspNet.FileProviders.Physical/FileSystemInfoHelper.cs
