+++ src/Microsoft.AspNet.FileProviders.Physical/project.json
@@ -1,7 +1,8 @@
 {
     "version": "1.0.0-*",
-    "description": "Implementation of ASP.NET 5 file provider abstractions.",
+    "description": "Implementation of ASP.NET 5 file provider abstractions for physical files.",
     "dependencies": {
+        "Microsoft.AspNet.FileProviders.Common": { "version": "1.0.0-*", "type": "build" },
         "Microsoft.AspNet.FileProviders.Interfaces": "1.0.0-*"
     },
     "frameworks": {
@@ -12,8 +13,6 @@
                 "System.Collections.Concurrent": "4.0.10-beta-*",
                 "System.IO.FileSystem.Watcher": "4.0.0-beta-*",
                 "System.IO.FileSystem": "4.0.0-beta-*",
-                "System.Linq": "4.0.0-beta-*",
-                "System.Reflection": "4.0.10-*",
                 "System.Runtime.Extensions": "4.0.10-beta-*",
                 "System.Text.RegularExpressions": "4.0.10-beta-*",
                 "System.Threading": "4.0.10-beta-*"
