+++ test/Microsoft.AspNet.FileProviders.Embedded.Tests/File.txt
