+++ test/Microsoft.AspNet.FileProviders.Embedded.Tests/Resources/File.txt
