+++ test/Microsoft.AspNet.FileProviders.Embedded.Tests/Resources/ResourcesInSubdirectory/File3.txt
