+++ test/Microsoft.AspNet.FileProviders.Embedded.Tests/sub/File2.txt
