+++ test/Microsoft.AspNet.FileProviders.Embedded.Tests/sub/dir/File3.txt
