+++ test/Microsoft.AspNet.FileProviders.Physical.Tests/File.txt
@@ -1,0 +1,1 @@
+﻿Hello
