+++ test/Microsoft.AspNet.FileProviders.Physical.Tests/project.json
@@ -1,6 +1,6 @@
 {
     "dependencies": {
-        "Microsoft.AspNet.FileProviders": "1.0.0-*",
+        "Microsoft.AspNet.FileProviders.Physical": "1.0.0-*",
         "Microsoft.AspNet.Testing": "1.0.0-*",
         "Shouldly": "1.1.1.1",
         "xunit.runner.aspnet": "2.0.0-aspnet-*"
@@ -8,7 +8,6 @@
     "commands": {
         "test": "xunit.runner.aspnet"
     },
-    "resources": "File.txt;sub/**;Resources/**",
     "frameworks": {
         "dnx451": { }
     }
