+++ test/Microsoft.AspNet.FileProviders.Physical.Tests/sub/File2.txt
@@ -1,0 +1,1 @@
+﻿Hello2
